fix: improved title positioning and font
