+++ nextjs/src/app/all/page.tsx
@@ -13,7 +13,7 @@ export default async function IndexPage() {
   );
 
   return (
-    <div className="container mx-auto max-w-7xl">
+    <div className="container mx-auto max-w-8xl">
       <ArtistGrid artists={artists} />
     </div>
   );
