+++ nextjs/src/app/producer-writers/page.tsx
@@ -2,6 +2,7 @@ import { client } from "@/sanity/client";
 import { ArtistGrid } from "@/components/artist-grid";
 import { type SanityDocument } from "next-sanity";
 import { QUERIES } from "@/constants/queries";
+import { Title } from "@/components/title";
 
 const options = { next: { revalidate: 30 } };
 
@@ -13,7 +14,8 @@ export default async function ProducerWriterPage() {
   );
 
   return (
-    <div className="container mx-auto max-w-7xl">
+    <div className="container mx-auto max-w-8xl">
+      <Title />
       <ArtistGrid artists={artists} />
     </div>
   );
